+++ enjoy/src/renderer/components/misc/index.ts
@@ -2,6 +2,9 @@ export * from "./db-state";
 export * from "./layout";
 export * from "./loader-spin";
 export * from "./login-form";
+export * from "./bandu-login-form";
+export * from "./email-login-form";
+export * from "./github-login-form";
 export * from "./no-records-found";
 export * from "./page-placeholder";
 export * from "./sidebar";
