+++ enjoy/src/renderer/components/misc/login-form.tsx
@@ -1,12 +1,10 @@
 import {
   Button,
   toast,
-  Input,
   Separator,
   Sheet,
   SheetContent,
   SheetTrigger,
-  Label,
   Card,
   CardContent,
   CardHeader,
@@ -19,12 +17,14 @@ import {
   UserSettings,
   LanguageSettings,
   LoaderSpin,
+  BanduLoginForm,
+  GithubLoginForm,
+  GithubLoginButton,
 } from "@renderer/components";
 import { ChevronLeftIcon } from "lucide-react";
-import intlTelInput from "intl-tel-input";
-import "intl-tel-input/build/css/intlTelInput.css";
 import { WEB_API_URLS } from "@/constants";
 import { useDebounce } from "@uidotdev/usehooks";
+import { EmailLoginForm } from "./email-login-form";
 
 export const LoginForm = () => {
   const { EnjoyApp, login, webApi, user } = useContext(
@@ -159,20 +159,7 @@ export const LoginForm = () => {
               {t("youCanAlsoLoginWith")}
             </div>
             <div className="flex items-center space-x-2 justify-center">
-              <Button
-                variant="outline"
-                size="icon"
-                data-tooltip-id="global-tooltip"
-                data-tooltip-content="GitHub"
-                className="w-10 h-10 rounded-full"
-                onClick={() => handleLogin("github")}
-              >
-                <img
-                  src="assets/github-mark.png"
-                  className="w-full h-full"
-                  alt="github-logo"
-                />
-              </Button>
+              <GithubLoginButton />
 
               <Button
                 variant="outline"
@@ -208,7 +195,7 @@ export const LoginForm = () => {
                 <SheetContent side="bottom" className="h-screen">
                   <div className="w-full h-full flex">
                     <div className="m-auto">
-                      <PandoLoginForm />
+                      <BanduLoginForm />
                     </div>
                   </div>
                 </SheetContent>
@@ -234,238 +221,5 @@ export const LoginForm = () => {
         </div>
       </div>
     </>
-  );
-};
-
-const EmailLoginForm = () => {
-  const [email, setEmail] = useState<string>("");
-  const [code, setCode] = useState<string>("");
-  const [codeSent, setCodeSent] = useState<boolean>(false);
-  const [countdown, setCountdown] = useState<number>(0);
-  const { login, webApi } = useContext(AppSettingsProviderContext);
-
-  useEffect(() => {
-    let timeout: NodeJS.Timeout;
-
-    if (countdown > 0) {
-      timeout = setTimeout(() => {
-        setCountdown(countdown - 1);
-      }, 1000);
-    }
-
-    return () => {
-      if (timeout) clearTimeout(timeout);
-    };
-  }, [countdown]);
-
-  return (
-    <div className="w-full">
-      <div className="w-full grid gap-4 mb-6">
-        <div className="grid gap-2">
-          <Label htmlFor="email">{t("email")}</Label>
-          <Input
-            id="email"
-            className="h-10"
-            type="email"
-            placeholder="m@example.com"
-            required
-            value={email}
-            disabled={countdown > 0}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        </div>
-
-        <div className="grid gap-2">
-          <Label htmlFor="code">{t("verificationCode")}</Label>
-          <Input
-            id="code"
-            className="h-10"
-            type="text"
-            required
-            minLength={5}
-            maxLength={5}
-            placeholder={t("verificationCode")}
-            value={code}
-            onChange={(e) => setCode(e.target.value)}
-          />
-        </div>
-      </div>
-
-      <div className="grid grid-cols-2 gap-4">
-        <Button
-          variant="secondary"
-          size="lg"
-          className="w-full"
-          disabled={!email || countdown > 0}
-          onClick={() => {
-            webApi
-              .loginCode({ email })
-              .then(() => {
-                toast.success(t("codeSent"));
-                setCodeSent(true);
-                setCountdown(120);
-              })
-              .catch((err) => {
-                toast.error(err.message);
-              });
-          }}
-        >
-          {countdown > 0 && <span className="mr-2">{countdown}</span>}
-          <span>{codeSent ? t("resend") : t("sendCode")}</span>
-        </Button>
-
-        <Button
-          variant="default"
-          size="lg"
-          className="w-full"
-          disabled={!code || code.length < 5 || !email}
-          onClick={() => {
-            webApi
-              .auth({ provider: "email", code, email })
-              .then((user) => {
-                if (user?.id && user?.accessToken) login(user);
-              })
-              .catch((err) => {
-                toast.error(err.message);
-              });
-          }}
-        >
-          {t("login")}
-        </Button>
-      </div>
-    </div>
-  );
-};
-
-const PandoLoginForm = () => {
-  const ref = useRef<HTMLInputElement>(null);
-  const [iti, setIti] = useState<any>(null);
-  const [phoneNumber, setPhoneNumber] = useState<string>("");
-  const [code, setCode] = useState<string>("");
-  const [codeSent, setCodeSent] = useState<boolean>(false);
-  const [countdown, setCountdown] = useState<number>(0);
-  const { login, webApi } = useContext(AppSettingsProviderContext);
-
-  const validatePhone = () => {
-    if (
-      iti?.isValidNumber() &&
-      iti?.getNumberType() === (intlTelInput.utils.numberType as any)?.MOBILE
-    ) {
-      setPhoneNumber(iti.getNumber());
-    } else {
-      setPhoneNumber("");
-    }
-  };
-
-  useEffect(() => {
-    if (!ref.current) return;
-
-    intlTelInput(ref.current, {
-      initialCountry: "cn",
-      utilsScript:
-        "https://cdn.jsdelivr.net/npm/intl-tel-input@19.2.12/build/js/utils.js",
-    });
-    setIti(intlTelInput(ref.current));
-
-    return () => {
-      iti?.destroy();
-    };
-  }, [ref]);
-
-  useEffect(() => {
-    iti?.setCountry("cn");
-  }, [iti]);
-
-  useEffect(() => {
-    let timeout: NodeJS.Timeout;
-
-    if (countdown > 0) {
-      timeout = setTimeout(() => {
-        setCountdown(countdown - 1);
-      }, 1000);
-    }
-
-    return () => {
-      if (timeout) clearTimeout(timeout);
-    };
-  }, [countdown]);
-
-  return (
-    <div>
-      <div className="flex items-center justify-center mb-4">
-        <img src="assets/bandu-logo.svg" className="w-20 h-20" alt="bandu" />
-      </div>
-
-      <div className="grid gap-6">
-        <div className="grid gap-4">
-          <div className="grid gap-2">
-            <Label htmlFor="phone">{t("phoneNumber")}</Label>
-            <input
-              id="phone"
-              value={phoneNumber}
-              onInput={validatePhone}
-              onBlur={validatePhone}
-              className="border text-lg py-2 px-4 rounded"
-              ref={ref}
-            />
-          </div>
-          <div className="grid gap-2">
-            <Label htmlFor="verrificationCode">{t("verificationCode")}</Label>
-            <Input
-              id="verrificationCode"
-              className="border py-2 h-10 px-4 rounded"
-              type="text"
-              minLength={5}
-              maxLength={5}
-              placeholder={t("verificationCode")}
-              value={code}
-              onChange={(e) => setCode(e.target.value)}
-            />
-          </div>
-        </div>
-
-        <div className="grid grid-cols-2 gap-4">
-          <Button
-            variant="secondary"
-            size="lg"
-            className="w-full"
-            disabled={!phoneNumber || countdown > 0}
-            onClick={() => {
-              webApi
-                .loginCode({ phoneNumber })
-                .then(() => {
-                  toast.success(t("codeSent"));
-                  setCodeSent(true);
-                  setCountdown(120);
-                })
-                .catch((err) => {
-                  toast.error(err.message);
-                });
-            }}
-          >
-            {countdown > 0 && <span className="mr-2">{countdown}</span>}
-            <span>{codeSent ? t("resend") : t("sendCode")}</span>
-          </Button>
-          <Button
-            variant="default"
-            size="lg"
-            className="w-full"
-            disabled={!code || code.length < 5 || !phoneNumber}
-            onClick={() => {
-              webApi
-                .auth({ provider: "bandu", code, phoneNumber })
-                .then((user) => {
-                  if (user?.id && user?.accessToken) login(user);
-                })
-                .catch((err) => {
-                  toast.error(err.message);
-                });
-            }}
-          >
-            {t("login")}
-          </Button>
-        </div>
-      </div>
-    </div>
   );
 };
