+++ enjoy/src/renderer/components/ui/sheet.tsx
@@ -72,7 +72,7 @@ const SheetContent = React.forwardRef<
       >
         {children}
         {displayClose && (
-          <SheetPrimitive.Close className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-secondary">
+          <SheetPrimitive.Close className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none disabled:pointer-events-none data-[state=open]:bg-secondary">
             <Cross2Icon className="h-4 w-4" />
             <span className="sr-only">Close</span>
           </SheetPrimitive.Close>
